Update deployment-guide-enrollment-macos.md
making changes to fix that - User affinity is established when you land on the home screen after the setup screens, however the device will not be fully registered with AAD until the Company Portal login. The device will not show up in a given user's device list in the AAD portal until the Company Portal login.