Update vpn-settings-configure.md
When creating VPN profiles the same constraint wrt to cert based auth and using the same group to deploy the profile AND associated cert profiles applies. We describe this as a TIP in the WIFI profile but not here in VPN - customers make this mistake in the same way so would be good to have consistent message.