Adding {{DeviceId}}
We added the ability to include the Intune device ID as a token in SCEP and PKCS certificate profiles about the same time we add URI as a SAN attribute; this should have been added to the docs at that time and was missed. This information will be very important as NAC vendors move to the new NAC service (Compliance Retrieval service).