Update resize-cloud-pc.md
Deleted the roles that don't have permissions to perform a resize. We havent released our AAD role (Windows 365 Admin) so it can't be documented and Cloud PC Admin is not an AAD role nor does it have UX permissions to access the MEM Devices list by itself, so would have to put a qualifier on that role.