Include instructions for converting AABs to APKs
The Intune App Wrapping Tool currently does not support AAB inputs or outputs. However, Google's `bundletool` can be used to convert an AAB to a single APK for use with the Wrapping Tool. Additionally, as of August 2021, private apps can continue to be uploaded to the Google Play Store as APKs.

Add a note with context about AABs to the prerequisite section and include instructions for converting an AAB to an APK for use with the Wrapping Tool.