Updates to the Security Tab on NDES certs
There is a vulnerability in re-using web server and other templates to create the NDES certificates. If the mspki-certificate-name-flag property is a 
bitmask and if the CT_FLAG_ENROLLEE_SUPPLIES_SUBJECT flag is present, a requester can specify the SAN and impersonate any user (domain admins) and request a certificate as them. 

"Supply in the Request" is a requirement for this cert to work, so authenticated users should be removed to reduce the attack surface of this vulnerability. 

(screenshots in the document show authenticated users as part of the security on these certificates)