Update names of Windows/Microsoft antivirus feature
Updated the names of the Windows/Microsoft antivirus feature as it appears in Server Manager. For some reason they keep changing the name of the feature. This lead to problems in one case where I was trying to verify if a customer had the Windows Defender Antivirus feature enabled in Windows Server 2022 as part of the prerequisite to the Endpoint Protection point and I could not find it in Server Manager. I eventually discovered it had been renamed to MICROSOFT Antivirus feature in Window Server 2022.