Update deploy-clients-to-windows-computers.md
Added clarification regarding certificate removal requirements to the "Prepare the client computer for imaging" section as we are finding that customers aren't necessarily removing the certificates from the SMS certificate store. These certificates are also used for client identity and can cause issues if not removed prior to image capture.