Update enrollment-autopilot.md
Change :  - To create a group that includes all of your Autopilot devices, enter: `(device.devicePhysicalIDs -any (_ -contains "[ZTDId]"))`.  
To:  - To create a group that includes all of your Autopilot devices, enter: `(device.devicePhysicalIDs -any (_ -contains "[ZTDID]"))`.
The parameter is case sensitive and will not work correctly if the case is not correct within the Azure AD group dynamic rule.