Revising URL to remove HTTPS formatting.
TIP FOR LINKS: Links to docs.microsoft.com content cannot publish with an HTTPS path. Instead, we can use a relative link (as done here) for a .md file in the same repository, or a path to the product/file/article, which would in this case have been (/intune/apps/intune-management-extension.md).   One other consideration (which isn't an issue here) is we never include the language path of the URL (like en-us). The absence of a language path ensures the reader will be linked to a content version that matches their browsers default language (if  that content version is available), and if it's not available, they are linked to the default of English.