Editing link
@rogersoMS.   Thanks for this PR to the public version of the docs repository. I've updated the link, which was in the doc as https://docs.microsoft.com/azure/active-directory/devices/assign-local-admin#:~:text=Administrator%20privileges%20using%20both%20policies%20are%20evaluated%20only%20for%20the%20following%20well%2Dknown%20groups%20on%20a%20Windows%2010%20device%20%2D%20Administrators%2C%20Users%2C%20Guests%2C%20Power%20Users%2C%20Remote%20Desktop%20Users%20and%20Remote%20Management%20Users. That seems to have picked up some junk that's not uncommon with many copy/paste actions for URLs.

The correct format for links to other docs.microsoft.com content requires us to not use an HTTPS url, removal of the language path  like en-us (which isn't an issue here), and use of a relative path of some sort to the destination. As this was to a different docs repository from memdocs, the link path for the subsection is  /azure/active-directory/devices/assign-local-admin#manage-administrator-privileges-using-azure-ad-groups-preview.