Removing (incorrect) TIP regarding silent enablement & Endpoint Security
Green TIP referring to BitLocker silent enablement not being possible with an Endpoint Security policy is incorrect. Both an Endpoint Security profile and a Device Configuration profile can achieve BitLocker silent enablement if the correct settings (now correted in this same doc) are used. We have tested both profiles and they allow a standard user to BitLocker encrypt the OS disk silently. Removing TIP as it appears to be technically incorrect.