Update help-support-windows-cpapp.md
Contribution provides help for Intune admins and support team, which isn't the audience of this article. However, I recognize that due to SEO and lots of keywords in this article, search results may lead IT pros to this article first. Since this is an article about help and support, I went ahead and created a new section at end of article for IT pro and admin resources,