adding AAD user information
Step 6 of https://docs.microsoft.com/en-us/mem/intune/protect/certificate-connector-install#configure-the-certificate-connector mentions On the Azure AD Sign In page, select the environment that hosts your Azure Active Directory, and then select Sign In. You’ll then be asked to authenticate your access. This user account must be a Global Admin or an Intune Admin with an Intune license assigned and the user must be a synchronized account from your local Active Directory.
That information isn't currently listed in the pre-requisites. I have received a few support cases where the customer called out that requirement wasn't listed in the pre-requisites.