Add Warning against placing console extension assemblies in bin folder.
Spend several days troubleshooting exactly this.  I realize that this whole process is now deprecated but I assure you that other vendors will be waaaaay behind the curve here.  Suggesting using the bin folder was really not a great suggestion/design in the first place and when I worked with the team on the new design I pushed hard to avoid this problem and they did.
If console extension A has dependency Y and so does extension B ... then unless their the _exact_ save version (spoiler: they won't be) then the last extension wins and the first is broken.