Added Pull-DP to DP configured as source
Even though the client on the Pull-DP is doing the download and there is a rule for client --> DP already it might make sense to also mention Pull-DP to Pull-Source as an extra rule. Especially when the Pull-DPs client would normally not reach the source DP due to a firewall in between.