Consider changing the word "must" to "should"
Setting the Feature Update deferral value to 0 in the Update ring settings is favourable and avoids unnecessary delays or complexities when targeting the same device with a Feature Update Policy. If the device is targeted with a Feature Update Policy and the Feature Update Deferral value in the Update Ring is set to a value higher than 0 then the feature update won't be offered to the device until the deferral period (in days) has been reached.

It is my understanding that the value does not "have" to be 0 but it is best practice to set this to 0. The use of the word "must" is ambiguous and suggests that the value must be zero for the Feature Update officered to the device, via the Feature Update policy, to be effective.

The Tip below the suggested change goes part way to explain the above to the reader