Updating added links
The links were pointing to a GitHub location and not a published docs location. That can be identified when you see /blob/ in the URL.  TO fix them, using the published doc location we can use the direct URL by starting with the /windows/ part of the URL.  

We only need the .md as part of a link when using a relative link to content in the same repository (such as from one Intune article to another).  These should now work (which I'll test before signing off).