Correct spelling and formatting
https://microsoft-ce-csi.acrolinx.cloud/api/v1/checking/scorecards/3576cc95-2103-4939-a969-ea386fa5db83#CORRECTNESS

Line 209: One of the following privileges are called: (are --> is)
Line 525: replace italic text with note formatting for "A setting of Enabled prevents slide shows from running."
Line 883: Revise "If the server supports Windows NT Challenge Response, the user's sign in uses the user's network user name and password for sign in." to "If the server supports Windows NT Challenge Response, the user's network user name and password are used for sign-in." Even better might be "If the server supports Windows NT Challenge Response, the server uses the user's network user name and password for sign-in." Unable to confirm the server is the correct subject, however.
Line 1117: capitalize "Java"
Line 1126: Internet Explorer include all network paths. (includes) However, this line duplicates the preceding line but without the formatting. Delete the duplicate line?
Line 1663: Add period after "executables" in "Zero tolerance blocks all unknown executables"
Line 1808: "If not, (and if the user has specified never to ask)," (delete parentheses, keep commas)