Fix typos
Line 25: ... when a user has create and delegate rights for an object type and wants to create an object and allow members of a user group to see it. --> ...when a user has created and delegated rights for an object type and wants to create an object and allow members of a user group to see it. 

Line 44: If the user only has instance-level read rights to certain collections, the user only has rights to see resources that are a members of those collections. (are a members --> are members)