Added-note-under-device-user-scope
Added an important alert under user or device scope that was suggested by Rui Zhao. 

He said "As that doc is for all Windows devices, we should keep the existing "User scope or Device scope" section as it is. While we add a !IMPORTANT alter for Holographic devices."

I asked if we should change other parts of the doc like the "Authentication method" section, but he said: 

[9:43 AM] Rui Zhao
No, we don't need to change "authentication method" as even it's a device scope VPN profile shared by multiple users the required certificate etc. would be per-user. 
 like 1

[9:45 AM] Rui Zhao
This is by-design. HL2/WCOS inherits the VPN design from Windows phone. It'd take some efforts to change it to Windows model and VPN team doesn't think it'd happen unless we have strong business justifications.

[9:45 AM] Rui Zhao
Your PR looks good to me.