Update device-enrollment-manager-enroll.md
Update two important pieces of information as after we cross the 1000 devices mark, if we add one device then it deletes another. Reported by multiple users as well. So the total number is 1000 I presume, we can check with PG if need be.

Similarly, 150 active DEM accounts per tenant is also updated.