Update microsoft-tunnel-prerequisites.md
We received a query asking if we support RHEL8.6. Our position is that we do not state support of any additional versions except the ones listed in the Supported Linux Distribution Table until we can prioritize and test support for a specific version.  

RHEL version support statement must clearly call out that we support only the versions stated in the Supported Linux Distributions table.