Update manage-microsoft-edge.md
https://microsoft-ce-csi.acrolinx.cloud/api/v1/checking/scorecards/ed6b93ab-a331-4f6e-9c76-3086b6d4c81d#CORRECTNESS
Line 178: If you have configured a homepage shortcut, that will override the first top site configured. --> If you have configured a homepage shortcut, that shortcut will override the first top site configured.
Line 316: Organization --> Organizations
Line 477: The layers of the network architecture is called the network stack. --> The layers of the network architecture are called the network stack.