Code sample expanded upon
The sample discovery script to me anyway, seemed short on a practical example that a customer could build upon, and made it harder to understand at first glance, what the script should actually be doing.

I've suggested using some code to first gather the results from the device, and then feed it into the hash, as I think it would help understanding. I've also removed the line in the output script that referenced a Microsoft employee's username - I don't think it is necessary to show the script being run, but just the actual output.