Update Exception Information for D4C
Information related to Defender for Cloud - Defender for Servers functionality being unavailable is incorrect. This dependency is linked to the agent leverage in Defender for Cloud - Defender for Servers (Unified Solution Agent), This is the default option in Defender for Servers P1 and has now been recently added to Defender for Servers P2 (Our original plan) - https://techcommunity.microsoft.com/t5/microsoft-defender-for-cloud/defender-for-servers-plan-2-now-integrates-with-mde-unified/ba-p/3527534

I have updated this exception to mention the SKU of Defender for Cloud and Agent Type which doesn't allowed for this functionality. Please feel free to change but it needs to be correct, I have had two partners reach out to me regarding this.