Attempt to reduce confusion in Wipe section
Customers have queried the table listed under the "Wipe device, but keep enrollment state and associated user account" option. It is not clear that the table shows what data would be retained "only" if that option is selected, and it is sometimes interpreted as a list of how wipe would work in general.