Correcting Connect-MSGraphApp to Connect-MSGraph
This is a reversal of change #2460 on 18th Feb 2022: https://github.com/MicrosoftDocs/memdocs/issues/2460
That change broke these steps for myself and my customer:
"Update-MSGraphEnvironment : Cannot validate argument on parameter 'AppId'. The argument is null or empty. Provide an argument that is not null or empty, and then try the command again.
Both of us have actually tested these steps (Connect-MSGraph) and ensured we can successfully progress to the next step.