Update Show/Hide Device to Device Migration
Both "Show" and "Hide" "Device to Device Migration" have no effect on destination devices with iOS/iPadOS 13.0 and later that are already enrolled with Apple. The option will not be shown. If this information were published here, it may save untold hours of frustration trying to make the option show up 😃  

Sources
 - https://www.reddit.com/r/Intune/comments/dfj8vv/has_anyone_gotten_device_to_device_migration_to/
 - https://developer.apple.com/forums/thread/123433
 - https://social.technet.microsoft.com/Forums/en-US/9faf8e23-3bfe-4b0e-90d9-51bf2a3c40b0/device-to-device-migration?forum=microsoftintuneprod
 - Personal experience

Regards,
Urbs3w