Updating PR.
Moved message into a Tip in the Tool usage section, and removed callout of the Bug. Reformatted the messaging a bit as well. 
Also revised earlier text to note tha the tool is in fact a PowerShell script, to help build relevance for the subsequent Tip that notes what it does and how to fix it.