Update whats-new-archive.md
Lines 116, 1090 Undo changes from corporate-owned to corporate owned. (This is a situation in which "personally-owned" devices is non-standard because "personally" is an adverb that modifies the verb "owned," but "corporate" (not corporately") is an adjective that combines with owned to modify devices, and thus requires the hyphen. If "corporately" had been used, the items would use a parallel structure and not use different punctuation.