Adding new common issues
Result of several support cases. I have added the markdown code for a picture of an error, but am unable to upload the picture file to the repo. Contact alias jascott to get the picture to upload please BEFORE publishing to prod. Feel free to update the formatting of the list/headings/etc.