Adding an additional setting for silent enablement
As per the tool tip in the Intune console, we also require the setting Require Key File Creation to be set to Allowed or Blocked for silent encryption. The verbiage from the tooltip is below for reference:

Recovery key file creation
Setting this to Allow will allow an admin user to create a 256-bit recovery key file manually. Setting this to Require will require an admin user to create a 256-bit recovery key file manually. Setting this to Deny will disallow any admin user from creating a 256-bit recovery key file manually. BitLocker silent enablement requires this setting configured as Deny or Allow.