Update Registry keys section
While it points to pre-provisioning scenario, the restart the device performs is what creates the condition needed for the issue. (In pre-provisioning mode, the device shutdown after technical flow and started to continue with the user flow.). So, this conflict affects any Autopilot scenario as long as a restart happens during the device ESP.