Update restore-overview.md
As observed by implementing the produce, the default "long-term" restore points are taken every 12 hours and removed after 7 days, for a totals of 13 restore points.  Change the name of "short-term" and "long-term".  A "short-term" restore point can be set for every 24 hours for a total of 10 restore points over 10 days, while "long-term" restore points only last for a total of 7 days, 3 days less then is possible with "short-term" restore points.