Update device-compliance-get-started.md
I’ve been working on a Strategic support case (33602944) with BP (British Petroleum), they this certain addition they want to incorporate. 
They want to add something along the lines of: “If an Enhanced jailbreak detection evaluation does not run for a certain period of time, the device will be marked as Jailbroken, and subsequently as Not Compliant.”
We’ve validated this to be correct from a technical standpoint and I could confirm this as the expected behavior for us. However, they are making the argument that the addition is needed in order to provide more clarity regarding the expected behavior. I can provide more details about the impact and the reasoning behind this request as/if needed.
Do you consider this to be feasible? I’ve made my argument to them regarding this being “self-explanatory” – we don’t document logical outcomes. Nevertheless, they seem to still want to add it, for full transparency purposes.