Revision
We should not remove the list of platforms from this location. That creates a required link out to another doc library to understand which platforms support this specific solution. Instead, we can add a link to that destination for those who require (at that time) the additional details about platform specific requirements for Defender.
 