Update requirements.md
Brazil South is a restricted region for Windows 365 and shouldn't be in the public docs. Its behind a deal desk exception. Because it showed up in the public docs, some customers are filing support tickets to get access to the region but they need to first go to deal desk and get an exception approved by them. Hence, the edit.