Clarify VPN support
The current wording for VPN support seems to indicate that Microsoft directly supports the VPN clients listed. Changed wording to indicate that the listed VPN clients have been tested, but support should be directed to the VPN vendors if there are any issues with the VPN client. Also clarified that this is not a comprehensive list of VPN clients that may work with Autopilot.