Editing for consistency with other entries in this article.
Removing the code-box formatting (which is a possible fix) so we can use the \ as an escape character, which should allow the entries to appear as intended.  Thanks for this PR as this display issue isn't one that's easy to catch visually (clearly, as it was published some time back).