Update whats-new.md
Adding "existing" to make it clear that the await final config setting applies to both new and existing enrollment profiles. 
 iOS/iPadOS 13+ devices enrolling with Setup Assistant with modern authentication   
* iOS/iPadOS 13+ devices enrolling without user affinity    
* iOS/iPadOS 13+ devices enrolling with Azure AD shared mode  
You will be able to enable or disable this feature in new and existing enrollment profiles with the new setting. We recommend you set it to Yes to utilizing the Awaiting final configuration feature during Setup Assistant. The default for existing enrollment profiles is No, and the default for new enrollment profiles is Yes.