Update vpn-settings-ios.md
We have an ICM Incident-396200589 Details - (https://portal.microsofticm.com/imp/v3/incidents/details/396200589/home), where customer is trying to upload exclude url list for zscaler VPN profile, but was able to upload only 50. This is by design. We need to update the documentation for this. The ICM has the details.