Adding a pointer to WebCP
For a compliance remediation message, we need to point users to instructions to sync with Windows; this doc describes how to this in Company Portal app, but since users might not have the CP app installed, we should also point them to the Intune Company Portal web app (a.k.a. "WebCP"). I've added a note and a link to the doc that describes how to do syncs from WebCP.