Remove unnecessary bullet
This bullet appears to be the cause for 16K of content to be rendered in one bullet under "Notices". By removing it, I hope to see the bottom of this page look normal. Before removal, every heading was indented as part of the list item, and every subsequent bullet was a hollow second-level bullet.