Update autopilot-faq.yml
Make more the scenario more clear. The issue is that applications will not install after ESP finished because user ESP is not processed and a device token is not generated until it is. So either going through the user esp or skipping it is required for the device token to be issued.