Update filters-device-properties.md
The statement "Devices must be Intune enrolled to use this app property" doesn't make sense and it's adding confusion to customer's reading this article. This filter property can be indeed used to filter a policy for a MAM scenario. The example for this property even reflects this:  '(app.deviceManagementType -eq "Unmanaged")'.