updating link to Apple Docs for passcode maxFailedAttempts
the article stating the following was moved by Apple so we were pointing to the wrong resource 

https://support.apple.com/en-in/guide/security/sec20230a10d/web 
Consecutive attempts of the same incorrect passcode don’t count towards the limit. This setting is also available as an administrative policy through a mobile device management (MDM) solution that supports this feature and through Microsoft Exchange ActiveSync, and can be set to a lower threshold.