Update app-discovered-apps.md
There may be a better way to word this but the IME uses a specific WMI class to scan for "Detected Apps" data. This is NOT limited to just MSI based apps. EXE based apps are also inventoried if they have a DIsplayName and DisplayVersion entry in the ARP registry keys