further clarifying target local time changes with 2401
the current working is confusing and important details are left out:

Symptom:
Starting with Intune service Release 2401, iOS/iPadOS DDM Software updates "Target Local Date Time:" setting is converting the time entered in the policy, to UTC.

Cause:
This is the expected behavior with the "Time and Sate" picker. The date time picker uses the browsers local time for reference and then converts it to UTC, so the Intune admin configuring the policy is in Eastern, it'll convert to UTC which is eastern +5 hours. However, that is the only time conversion that happens though, whatever the result is, it'll still deliver to that time to every time zone. For instance:
	• For an admin located in Eastern US, if the policy is configured with:
		○ 12:30 PM
The conversion will be UTC +5 (5:30 PM), now every device targeted by that policy will get have the deadline for 5:30 PM (whatever the local time for the device is). So a device with a local Australia time will still see the deadline for their 5:30 PM.