Update intune-endpoints.md to include ECS and fwlink domain
Updating the public lists of endpoints that Intune clients hit so as to include config.edge.skype.com and go.microsoft.com.  The former is used by clients to deliver new features.  The latter is the fwlink domain that is used to redirect traffic to other endpoints.  This control over redirection allows us to not break old clients that have static URI's baked into their bits when a URI needs to change.  Primary use cases of fwlinks in clients are around endpoint discovery and linking customers to help documents.