Update remove-central-administration-site.md
Adjusted the wording on the post-setup tasks item regarding updating Boot Images and Boot Media due to the trusted root key changing from the CAS being removed. The new wording is more in line with the other line items and removes any ambiguity as to whether or not the actions need to be taken.