Change additional to more than one
For command -line arguments, changed "additional" to "more than one". It's not clear what is meant by additional. Additional to what? When testing line-of-business apps I added multiple command line arguments which was causing the app to fail. When I looked at this doc, I saw the mention of "additional command-line options", but it wasn't clear what it was in addition to. I deduced that it was in addition to the first command-line argument, so I removed the "additional" command-line argument and it fixed the problem. However, customers may not make the same assumption and may not understand why they app isn't working. I've reworded to make it clearer that only command-line argument is allowed.