Add images from "April 2017" in whats-new-app-ui.md
In whats-new-app-ui.md, under the heading "April 2017", there are two subheadings—"New icons for the Managed Browser and the Company Portal" and "Sign in progress indicator in Android Company Portal"—that have site-relative links to images. The paths of those look like they'd be in memdocs-pr: /mem/intune/media. However, I can't find any of these images in the repo. I'm concerned about renaming paths still somehow affecting these images, so I've downloaded them to add them in the folder for whats-new-app-ui.md, which is memdocs-pr/intune/intune-service/fundamentals/media/whats-new-app-ui/.

Also, I corrected one of the images under "New icons". Both image references in that section had "before" in their file names, despite being labeled "Before" and "After". I changed the file name of the second to replace "before" with "after", and now the images are different.

The situation in "Sign in progress" is similar in that the images are supposed to show before and after, but both image references use the same file name (and it doesn't have "before" or "after" in its name).