Update mde-security-integration.md
I’ve finally been able to sync with PM/Dev on this. To preserve your contribution to content,  I’ve submitted edits to your PR to round it out. Thanks for submitting this PR as its raised visibility to something that I was unaware of (visibility in Defender) and for which no content asks were passed to me by PM/Dev.

The ongoing plan for this content is to:  
-  Continue listing  only those Intune profiles that are visible in both Intune admin center and in the Defender security console.  **Device control** is missing form that list, and was submitted by you via this content PR. 

-  For those Intune endpoint security profiles visible in the Defender console that are not supported by the Microsoft Defender security settings management  scenario (mde-attach), we’ll use a common note to call out the exception. That note exists for **Windows Security Experience**, and it will be modified to be generic for any profile. 

As part of the PM/Dev sync, I’ve been informed that we should expect to see additional Intune endpoint security profiles appear in Defender – and most or all should onboarded to the MDE-attach scenario at some point.  As new profiles are brought to my attention as visible, we’ll add them to the list. Unless they are also supported by mde-attach, we’ll add the note about mde-attach exceptions, as for *Windows Security Experience* today.