Update azure-virtual-desktop-multi-session.md
:) Needs PM review and verification.

Filed based on investigations from 2 customer incidents, in which clean-up rules are not honored. 

Need to verify ok to link to blog, as that's only thing that mentions "soft delete". 

If not wanting to mention "soft delete" and/or link to blog, can just change to deleted automatically after 30 days once confirmed, and link to core docs instead: https://learn.microsoft.com/en-us/mem/intune/remote-actions/devices-wipe#automatically-delete-devices-with-cleanup-rules, which is linked off of blog.