Update mde-security-integration.md
Changing presentation, as Amazon version branding might not support the 'or later' we see for other platforms. 

This will help avoid confusion,  as it seems AL2 is based on CentOS/RHEL 7, while AL2023 is based on Fedora/CentOS Stream - which seems to be a significant version difference worth a seperate version bullet point.