Update overview.md
Updating overview doc to clarify that the Windows of version of the client needs to be an Intune supported version of Windows for comanagement to be available. Based on multiple support cases from customers thinking that tenant attach and comanagment are the same thing, but they are not. Trying to help make that more clear.