Add macOS agent release notes to Intune Documentation
I'm PM on Apple Intune feature team. We previously did not have release notes published for macOS agent (which is a part under macOS MDM). I'm not 100% sure where's the best to add this new info. Maybe a new page? I'm adding to "What's New" for now and appreciate to discuss with documentation team on what's best presentation. Thank you.