Update encrypt-devices-filevault.md
If the Defer setting doesn't turn on "true",  FileVault setting is unable to be applied to MacOS devices normally.
Endpoint Protection policy settings for FileVault descripts "Defer the prompt until the user signs out. Only 'yes' is supported", and it wrote same in the below workitem. Defer setting must be enabled if use Setting Catalogs also.

https://msazure.visualstudio.com/Intune/_workitems/edit/7311006
In 2001 (https://msazure.visualstudio.com/Intune/_workitems/edit/5888687) we deleted this setting from the UI because only the 'yes' value, setting the boolean to 'true', is supported. The NC value of 'false' is not supported because Intune is not able to actually enforce that value and not defer until the user signs out (there might be an Apple bug we filed that i can't find at the moment). The user will always have deferral until they sign out that first time.