Update FileVault settings configuration steps
Thank you for this PR and the details about it.  

In looking at the note, I am going to lightly edit it for emphasis but also move it to where I will add direction to set Defer as part of Step 5, which is the step where the settings catalog File Vault settings are selected. 

If directing the customer to set Defer is not something they should always do, let me know and I will remove the implied configuration that I'm adding to that step but would leave the note to maintain the reference at that point in the procedure.