issue: device-profile-assign.md, line 122
This isn't vandalism, I just do not see a way to report inaccurate/missing doc, and i do not know the answer. Line 121 specifically talks about devices  with no associated user and "device groups" - but does not make it clear how to do this. With iOS at least, not having a user associated means the  device does not show in Entra, and cannot be added to a group - even if the device shows in InTune.