Update wi-fi-settings-android-enterprise.md
Included the content in the Note Section

When using any EAP type (EAP-TLS, PEAP, or EAP-TTLS) and certificates for authentication, it is required to include the user principal name (UPN) in the Subject Alternative Name (SAN) for user or device certificates. If the UPN is not present in the SAN, the Wi-Fi profile deployment will fail.