Update app-protection-policy-settings-android.md
Modified the description to clarify that the setting pertains to the behavior when selecting a link for messaging apps (e.g., sms:xxxxxxxxxx), as it overlapped with the content of the [Transfer telecommunications data to] setting. Also, removed one instance of the sentence "For this setting, choose how to handle this type of content transfer when it's initiated from a policy-managed app," as it was repeated twice.